refactor(server): refactor actions
separate actions into their own files
